added new parameter "download-manifest"
The new parameter is required to give download_url a list of urls to be
downloaded.

Currently if the service download_url is running in service container, for
each url a service has to be defined. This means on the source service
server a container has to be spawned for each url.
In our current infrastructure it takes about 5sec for each url/container.
In a project with over 1300 defined urls this leads to a timeout from the
service dispatcher.

With this patch you could give the service a list urls and the container has
to be spawned only once.